feat: Implement HTML table view for Run data and update RunController
This commit includes changes to the 'mainscreen.html' file to display Run data in a table format. It also updates the 'RunController.java' file to handle the GET request for the main screen and populate the model with Run data from the RunRepository. The controller has been switched from RestController to Controller to support Thymeleaf HTML views.